flyer_requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
frozenlist==1.3.3
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.3
multidict==6.0.4
numpy==1.24.1
openai==0.26.1
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.2
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
tqdm==4.64.1
urllib3==1.26.14
yarl==1.8.2
